+++ src/Container.cc
@@ -1,9 +1,9 @@
 #include "Container.h"
 
-size_t Container::s_capacity = 256;
-int Container::s_prefix_length = 1;
-int Container::s_prfx_prefix_length = 1;
-int Container::s_prfx_suffix_length = 1;
+uint16_t Container::s_capacity = 256;
+uint8_t Container::s_prefix_length = 1;
+uint8_t Container::s_prfx_prefix_length = 1;
+uint8_t Container::s_prfx_suffix_length = 1;
